# Optimize 251 a tiny bit